+++ C/menor_numero.c
@@ -24,7 +24,7 @@ int main()
 float menor_numero(float numero[], int tamanho_vetor)
 {
     //calcula o menor número.
-    int menor = numero[0];
+    float menor = numero[0];
     for (int i = 1; i < tamanho_vetor; i++)
     {
         if (numero[i] < menor)
